Move debug logic into Intersect.js, added tap to disable
